requirements.txt
django>=4.2.2
gunicorn==20.1.0
whitenoise==6.5.0
python-dotenv==1.0.0
mysqlclient==2.2.0
dj-database-url==2.0.0
djangorestframework==3.14.0